Refactoring - Created Applications Reducer
- moved Action/Effect "createApplication"
