+++ src/app/mission-control/activator-store/activator-store.actions.ts
@@ -1,5 +1,4 @@
 import { createAction, props } from '@ngrx/store';
-import { Application } from '../applications/applications.model';
 
 export const setCategoriesCount = createAction(
   '[ActivatorStore] set-categories-count',
@@ -21,8 +20,3 @@ export const grantAccess = createAction(
   props<{ activatorId: number; teamId: string }>()
 );
 export const requestAccess = createAction('[ActivatorStore] request-access', props<{ id: number }>());
-
-export const createApplication = createAction(
-  '[ActivatorStore] create-application',
-  props<{ application: Application }>()
-);
