+++ src/app/mission-control/applications/applications-create/applications-create.component.ts
@@ -4,7 +4,8 @@ import { ActivatedRoute } from '@angular/router';
 import { Store, select } from '@ngrx/store';
 import { MatDialog } from '@angular/material';
 import { KeyValue } from '@angular/common';
-import { setProgress, createApplication } from '@app/mission-control/activator-store/activator-store.actions';
+import { createApplication } from '../applications.actions';
+import { setProgress } from '@app/mission-control/activator-store/activator-store.actions';
 import { Activator } from '@app/mission-control/activator-store/activator-store.model';
 import { Application } from '../applications.model';
 import { Solution } from '@app/mission-control/solutions/solutions.model';
