+++ src/app/mission-control/applications/applications.actions.ts
@@ -1,0 +1,7 @@
+import { createAction, props } from '@ngrx/store';
+import { Application } from './applications.model';
+
+export const createApplication = createAction(
+  '[Applications] create-application',
+  props<{ application: Application }>()
+);
