+++ src/app/mission-control/applications/applications.module.ts
@@ -1,6 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { SharedModule } from '@app/shared/shared.module';
+import { StoreModule } from '@ngrx/store';
+import { EffectsModule } from '@ngrx/effects';
+import { featureKey, reducer } from './applications.reducer';
+import { ApplicationsEffects } from './applications.effects';
 
 import { ApplicationsDeploymentsComponent } from './applications-deployments/applications-deployments.component';
 
@@ -15,6 +19,8 @@ import { MatTableModule } from '@angular/material/table';
   imports: [
     CommonModule,
     SharedModule,
+    StoreModule.forFeature(featureKey, reducer),
+    EffectsModule.forFeature([ApplicationsEffects]),
     MatFormFieldModule,
     MatInputModule,
     MatPaginatorModule,
